requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-decouple==3.8
drf-spectacular==0.27.1
django-filter==23.3
Pillow==10.1.0
celery==5.3.4
redis==5.0.1
django-extensions==3.2.3
django-debug-toolbar==4.2.0
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2
openpyxl==3.1.2
xlsxwriter==3.1.9
